scripts/classifier/requirements.txt
absl-py==0.8.0
asn1crypto==0.24.0
astor==0.8.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
conda==4.7.12
conda-package-handling==1.3.11
cryptography==2.7
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
libarchive-c==2.8
Markdown==3.1.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.3.0
numpy==1.16.5
Pillow==6.1.0
protobuf==3.9.2
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.0
requests==2.22.0
ruamel-yaml==0.15.46
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.32.1
urllib3==1.24.2
Werkzeug==0.16.0
wrapt==1.11.2
